Change basis functions for symmetric matrix

+++ core/common/eigen.hpp
@@ -49,14 +49,14 @@ using dynamic_matrix = Eigen::Matrix<T, Eigen::Dynamic, Eigen::Dynamic>;
 template<typename T>
 using dynamic_vector = Eigen::Matrix<T, Eigen::Dynamic, 1>;
 
-template<typename T, size_t M, size_t N>
+template<typename T, int M, int N>
 using static_matrix = Eigen::Matrix<T, M, N>;
 
-template<typename T, size_t N>
+template<typename T, int N>
 using static_vector = Eigen::Matrix<T, N, 1>;
 
 // to mimic fourth order tensor
-template<typename T, size_t N>
+template<typename T, int N>
 using static_tensor = Eigen::Matrix<T, N * N, N * N>;
 
 template<typename T, size_t M, size_t N>
